+++ Gemfile
@@ -5,3 +5,5 @@ gem "sinatra"
 gem "blurrily"
 gem "rspec"
 gem "babel-transpiler"
+gem "haversine"
+gem "postcodes_io"
